GOV-5391

+++ d360.core/entities/Metric/MetricCondition.cs
@@ -8,7 +8,7 @@ namespace d360.core.entities.Metric
     public class MetricCondition : BaseObject
     {
         [DataMember, Key, Column(Order = 1)]
-        public long MapID { get; set; }
+        public int MapID { get; set; }
 
         [DataMember, Key, Column(Order = 2)]
         public int FieldTypeID { get; set; }
